i2c auto reset: make auto reset timeout opt-out
Add a default auto reset timeout to the BMC app and always start the auto
reset timer.

Add a Kconfig option (TT_AUTO_RESET) to SMC to enable auto reset by default.
If this is enabled, then  telemetry heartbeat is periodically sent to BMFW every
telemetry_update_interval. If this is disabled, then telemetry heartbeat is
never sent and a message is sent to BMC during SMC init to disable the auto
reset timer.

Signed-off-by: Petra Alexson <palexson@tenstorrent.com>